Move export functions to preload
